Split the ProjectPlanning types out into another module
Not needed immediately, but we'll need it eventually to avoid import
cycles later. No other changes, just moving things around and pruning
imports and language extensions.
